Develop

+++ TaskManagerModel/DataModel.cs
@@ -45,6 +45,8 @@ namespace TaskManagerModel
 
         private readonly IContextFactory _contextFactory;
 
+        private readonly IValidator<UserTask> _validator; 
+
         public DataModel(IContextFactory contextFactory, ITaskFilter taskFilter)
         {
             _contextFactory = contextFactory;
@@ -52,6 +54,7 @@ namespace TaskManagerModel
             _taskFilter = taskFilter;
             _filterType = FilterType.All;
             _sortType = SortType.AscendingPriority;
+            _validator = new UserTaskValidator();
         }
 
         public void AddTask(UserTask task)
@@ -59,8 +62,7 @@ namespace TaskManagerModel
             if (task == null)
                 throw new ArgumentNullException(nameof(task));
 
-            var validator = new UserTaskValidator();
-            validator.ValidateAndThrow(task, ruleSet: "Body");
+            _validator.ValidateAndThrow(task, ruleSet: "Body");
 
             using (var context = _contextFactory.BuildContex())
             {
@@ -75,8 +77,7 @@ namespace TaskManagerModel
             if (task == null)
                 throw new ArgumentNullException(nameof(task));
 
-            var validator = new UserTaskValidator();
-            validator.ValidateAndThrow(task, ruleSet: "*");
+            _validator.ValidateAndThrow(task, ruleSet: "*");
 
             using (var context = _contextFactory.BuildContex())
             {
@@ -91,8 +92,7 @@ namespace TaskManagerModel
             if (task == null)
                 throw new ArgumentNullException(nameof(task));
 
-            var validator = new UserTaskValidator();
-            validator.ValidateAndThrow(task, ruleSet: "Id");
+            _validator.ValidateAndThrow(task, ruleSet: "Id");
 
             using (var context = _contextFactory.BuildContex())
             {
@@ -112,8 +112,10 @@ namespace TaskManagerModel
             {
                 foreach (var task in tasksList)
                 {
-                    var validator = new UserTaskValidator();
-                    validator.ValidateAndThrow(task, ruleSet: "Id");
+                    if(task == null)
+                        throw new ArgumentException(nameof(tasksList));
+
+                    _validator.ValidateAndThrow(task, ruleSet: "Id");
 
                     context.Delete(task);
                 }
@@ -150,6 +152,9 @@ namespace TaskManagerModel
         {
             List<UserTask> tasks;
             Expression<Func<UserTask, bool>> compare;
+
+            if(endDate < beginDate)
+                throw new ArgumentException($"{nameof(endDate)} < {nameof(beginDate)}");
 
             if (beginDate == endDate) //Вынести в поле класса?
                 compare = t => t.TaskDate == beginDate;
